% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_clm.R
\name{get_CL_vcov}
\alias{get_CL_vcov}
\title{Clustered standard errors for \code{clm}}
\usage{
get_CL_vcov(model, cluster)
}
\arguments{
\item{model}{clm object}

\item{cluster}{vector identifying clusters}
}
\value{
variance/covariance matrix
}
\description{
Clustered standard errors for \code{clm}
}
