% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_clm.R
\name{glance.clm}
\alias{glance.clm}
\title{Glance method for \code{clm} objects}
\usage{
glance.clm(x, ...)
}
\arguments{
\item{x}{clm object}

\item{...}{extra arguments}
}
\value{
data frame
}
\description{
Glance method for \code{clm} objects
}
\examples{
\dontrun{
#' require(ordinal)
fm1 <- clm(rating ~ temp * contact, data = wine)
glance(fm1)
}
}
