% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_clm.R
\name{tidy.clm}
\alias{tidy.clm}
\title{Tidy method for \code{clm} objects.}
\usage{
\method{tidy}{clm}(x, conf.int = FALSE, conf.level = 0.95,
  exponentiate = FALSE, quick = FALSE, cluster = FALSE, vcov = NULL,
  ...)
}
\arguments{
\item{x}{clm object}

\item{conf.int}{Logical: whether to include a confidence interval}

\item{conf.level}{confidence level of the interval, used only if \code{conf.int = TRUE}}

\item{exponentiate}{Logical: whether to exponentiate the coefficient estimates and confidence intervals}

\item{quick}{Logical: whether to compute a smaller and faster version}

\item{cluster}{Logical: whether to use clustered standard errors}

\item{vcov}{an alternative variance/covariance matrix, used only if \code{cluster = TRUE}}

\item{...}{extra arguments}
}
\value{
tidy data frame
}
\description{
This method tidies the coefficients of a \code{clm} object into a data frame.
}
\details{
If you have missing values in your model data, you may need to refit the model with
\code{na.action = na.exclude}.
}
\examples{
\dontrun{
require(ordinal)
fm1 <- clm(rating ~ temp * contact, data = wine)
tidy(fm1)
}
}
