% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_clm.R
\name{tidy.summary.clm}
\alias{tidy.summary.clm}
\title{Tidy method for summaries of \code{clm} objects}
\usage{
tidy.summary.clm(x, cluster = FALSE, vcov = NULL, digits = 2, ...)
}
\arguments{
\item{x}{clm object}

\item{cluster}{Logical: whether to calculate clustered standard errors}

\item{vcov}{alternative variance/covariance matrix, only used if \code{cluster = TRUE}}

\item{digits}{digits to print}

\item{...}{additional arguments}
}
\value{
tidy data frame
}
\description{
Tidy method for summaries of \code{clm} objects
}
\examples{
\dontrun{
#' require(ordinal)
fm1 <- clm(rating ~ temp * contact, data = wine)
tidy(summary(fm1))
}
}
